test(ir/ops): expand RMSNorm lowering and provider validation tests
Add comprehensive end-to-end tests for the RMSNorm IR op to verify
lowering correctness and provider implementation semantics. Changes
include:
- Verify provider registration and platform-specific availability
- Validate native implementation semantics (scaling, weight handling)
- Compare provider outputs against native reference for numerical accuracy
- Ensure dispatch matches direct implementation calls
- Test unsupported argument handling (e.g., variance_size override)

This ensures all IR lowering paths produce correct results and adhere
to expected native behavior across supported backends.
